+++ src/qcm/model/base.py
