+++ src/qcm/model/bdd_init.py
