+++ src/qcm/vue/editeur/editeur_callback_type.py
